Updated ReadMe
Spelled-out MAEC, OVAL, and CybOX in intro text. Replaced the links to three mitre.org sites with two github site link and CIS's url for OVAL. Added a link to the community discussion list in the feedback section. Added an About MAEC section with a link to the MAEC main website, and a Getting Help section with links for feedback and questions, to make this consistent with the other ReadMe pages in this format.